# lib/api-zod/src/generated/types/friendRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salli Islamic Prayer App API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialUser } from "./socialUser";

export interface FriendRequest {
  id: number;
  user: SocialUser;
  createdAt: Date;
}