# lib/api-zod/src/generated/types/notificationsCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salli Islamic Prayer App API
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationsCount {
  pendingRequests: number;
  unreadMessages: number;
}