# lib/api-zod/src/generated/types/sendDirectMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salli Islamic Prayer App API
 * OpenAPI spec version: 0.1.0
 */

export interface SendDirectMessageBody {
  /** @minLength 1 */
  content: string;
}