# lib/api-zod/src/generated/types/socialUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salli Islamic Prayer App API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialUserFriendshipStatus } from "./socialUserFriendshipStatus";

export interface SocialUser {
  id: number;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  city?: string | null;
  friendshipStatus: SocialUserFriendshipStatus;
}